requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI - OpenAI Agents
openai>=1.0.0
openai-agents-python @ git+https://github.com/openai/openai-agents-python.git

# Database
supabase>=2.3.4
postgrest>=0.16.0

# Config & Validation
python-dotenv==1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
httpx>=0.26.0

# Production
gunicorn==21.2.0